requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
autopep8==2.0.1
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
dj-database-url==1.0.0
Django==4.1.2
djangorestframework==3.14.0
gunicorn==20.1.0
kombu==5.2.4
numpy==1.23.4
opencv-python==4.6.0.66
Pillow==9.2.0
prompt-toolkit==3.0.36
psycopg2==2.9.5
pycodestyle==2.10.0
python-dotenv==0.21.0
pytz==2022.5
redis==4.5.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.5
uuid==1.30
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.2.0
